requirements.txt
Flask>=2.0
python-telegram-bot>=13.0
python-dotenv>=0.19
requests>=2.20 